+++ libft.h
@@ -13,12 +13,15 @@
 #ifndef LIBFT_H
 # define LIBFT_H
 
+#include <unistd.h>
+#include <stdlib.h>
+
 int	ft_isalpha(int c);
 int	ft_isdigit(int c);
 int	ft_isalnum(int c);
 int	ft_isascii(int c);
 int	ft_isprint(int c);
-unsigned int	ft_strlen(const char *str);
+size_t	ft_strlen(const char *str);
 void	*ft_memset(void *s, int c, unsigned int n);
 void	ft_bzero(void *s, unsigned int n);
 void	ft_bzero(void *s, unsigned int n);
